Adding caching to products and location setting

+++ server/account.go
@@ -4,13 +4,15 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/densestvoid/krogerrecipeshopper/app"
 	"github.com/densestvoid/krogerrecipeshopper/data"
+	"github.com/densestvoid/krogerrecipeshopper/kroger"
 	"github.com/densestvoid/krogerrecipeshopper/templates"
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 )
 
-func NewAccountMux(repo *data.Repository) func(r chi.Router) {
+func NewAccountMux(config Config, repo *data.Repository, cache *data.Cache) func(r chi.Router) {
 	return func(r chi.Router) {
 		r.Get("/", func(w http.ResponseWriter, r *http.Request) {
 			accountID, err := GetAccountIDFromRequestSessionCookie(repo, r)
@@ -31,7 +33,34 @@ func NewAccountMux(repo *data.Repository) func(r chi.Router) {
 				return
 			}
 
-			if err := templates.Account(account, profile).Render(w); err != nil {
+			authClient := kroger.NewAuthorizationClient(http.DefaultClient, kroger.PublicEnvironment, config.ClientID, config.ClientSecret)
+			authResp, err := authClient.PostToken(r.Context(), kroger.ClientCredentials{
+				Scope: kroger.ScopeProductCompact,
+			})
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			productsClient := kroger.NewProductsClient(http.DefaultClient, kroger.PublicEnvironment, authResp.AccessToken)
+			locationsClient := kroger.NewLocationsClient(http.DefaultClient, kroger.PublicEnvironment, authResp.AccessToken)
+
+			krogerManager := app.NewKrogerManager(productsClient, locationsClient, cache)
+
+			var location *data.CacheLocation
+			if account.LocationID != nil {
+				loc, err := krogerManager.GetLocation(r.Context(), *account.LocationID)
+				if err != nil {
+					http.Error(w, fmt.Sprintf("getting location: %v", err), http.StatusInternalServerError)
+					return
+				}
+				location = &loc
+			}
+
+			if err := templates.AccountPage(templates.Account{
+				ID:        account.ID,
+				ImageSize: account.ImageSize,
+				Location:  location,
+			}, profile).Render(w); err != nil {
 				http.Error(w, err.Error(), http.StatusInternalServerError)
 				return
 			}
@@ -72,18 +101,90 @@ func NewAccountMux(repo *data.Repository) func(r chi.Router) {
 			})
 
 			r.Patch("/settings", func(w http.ResponseWriter, r *http.Request) {
-				r.ParseForm()
-				accountID := uuid.MustParse(chi.URLParam(r, "accountID"))
-				imageSize := r.FormValue("imageSize")
-				if err := repo.UpdateAccountImageSize(r.Context(), accountID, imageSize); err != nil {
-					http.Error(w, fmt.Sprintf("updating account image size: %v", err), http.StatusInternalServerError)
+				callingAccountID, err := GetAccountIDFromRequestSessionCookie(repo, r)
+				if err != nil {
+					http.Error(w, fmt.Sprintf("getting account id: %v", err), http.StatusUnauthorized)
 					return
 				}
+				requestedAccountID := uuid.MustParse(chi.URLParam(r, "accountID"))
+				if callingAccountID != requestedAccountID {
+					http.Error(w, "deleting account", http.StatusUnauthorized)
+					return
+				}
+
+				r.ParseForm()
+				accountID := uuid.MustParse(chi.URLParam(r, "accountID"))
+
+				if r.Form.Has("imageSize") {
+					imageSize := r.FormValue("imageSize")
+					if err := repo.UpdateAccountImageSize(r.Context(), accountID, imageSize); err != nil {
+						http.Error(w, fmt.Sprintf("updating account image size: %v", err), http.StatusInternalServerError)
+						return
+					}
+				}
+
+				if r.Form.Has("locationID") {
+					locationID := r.FormValue("locationID")
+					var locationIDStr *string
+					if locationID != "" {
+						locationIDStr = &locationID
+					}
+					if err := repo.UpdateAccountLocationID(r.Context(), accountID, locationIDStr); err != nil {
+						http.Error(w, fmt.Sprintf("updating account location ID: %v", err), http.StatusInternalServerError)
+						return
+					}
+
+					account, err := repo.GetAccountByID(r.Context(), accountID)
+					if err != nil {
+						http.Error(w, fmt.Sprintf("getting account: %v", err), http.StatusInternalServerError)
+						return
+					}
+
+					authClient := kroger.NewAuthorizationClient(http.DefaultClient, kroger.PublicEnvironment, config.ClientID, config.ClientSecret)
+					authResp, err := authClient.PostToken(r.Context(), kroger.ClientCredentials{
+						Scope: kroger.ScopeProductCompact,
+					})
+					if err != nil {
+						http.Error(w, err.Error(), http.StatusInternalServerError)
+						return
+					}
+					productsClient := kroger.NewProductsClient(http.DefaultClient, kroger.PublicEnvironment, authResp.AccessToken)
+					locationsClient := kroger.NewLocationsClient(http.DefaultClient, kroger.PublicEnvironment, authResp.AccessToken)
+
+					krogerManager := app.NewKrogerManager(productsClient, locationsClient, cache)
+
+					var location *data.CacheLocation
+					if account.LocationID != nil {
+						loc, err := krogerManager.GetLocation(r.Context(), *account.LocationID)
+						if err != nil {
+							http.Error(w, fmt.Sprintf("getting location: %v", err), http.StatusInternalServerError)
+							return
+						}
+						location = &loc
+					}
+
+					if err := templates.LocationNode(location).Render(w); err != nil {
+						http.Error(w, err.Error(), http.StatusInternalServerError)
+						return
+					}
+				}
+
 				w.WriteHeader(http.StatusOK)
 			})
 
 			r.Route("/profile", func(r chi.Router) {
 				r.Get("/", func(w http.ResponseWriter, r *http.Request) {
+					callingAccountID, err := GetAccountIDFromRequestSessionCookie(repo, r)
+					if err != nil {
+						http.Error(w, fmt.Sprintf("getting account id: %v", err), http.StatusUnauthorized)
+						return
+					}
+					requestedAccountID := uuid.MustParse(chi.URLParam(r, "accountID"))
+					if callingAccountID != requestedAccountID {
+						http.Error(w, "deleting account", http.StatusUnauthorized)
+						return
+					}
+
 					accountID := uuid.MustParse(chi.URLParam(r, "accountID"))
 
 					account, err := repo.GetAccountByID(r.Context(), accountID)
@@ -98,7 +199,34 @@ func NewAccountMux(repo *data.Repository) func(r chi.Router) {
 						return
 					}
 
-					if err := templates.Profile(account, profile).Render(w); err != nil {
+					authClient := kroger.NewAuthorizationClient(http.DefaultClient, kroger.PublicEnvironment, config.ClientID, config.ClientSecret)
+					authResp, err := authClient.PostToken(r.Context(), kroger.ClientCredentials{
+						Scope: kroger.ScopeProductCompact,
+					})
+					if err != nil {
+						http.Error(w, err.Error(), http.StatusInternalServerError)
+						return
+					}
+					productsClient := kroger.NewProductsClient(http.DefaultClient, kroger.PublicEnvironment, authResp.AccessToken)
+					locationsClient := kroger.NewLocationsClient(http.DefaultClient, kroger.PublicEnvironment, authResp.AccessToken)
+
+					krogerManager := app.NewKrogerManager(productsClient, locationsClient, cache)
+
+					var location *data.CacheLocation
+					if account.LocationID != nil {
+						loc, err := krogerManager.GetLocation(r.Context(), *account.LocationID)
+						if err != nil {
+							http.Error(w, fmt.Sprintf("getting location: %v", err), http.StatusInternalServerError)
+							return
+						}
+						location = &loc
+					}
+
+					if err := templates.Profile(templates.Account{
+						ID:        account.ID,
+						ImageSize: account.ImageSize,
+						Location:  location,
+					}, profile).Render(w); err != nil {
 						http.Error(w, err.Error(), http.StatusInternalServerError)
 						return
 					}
@@ -106,6 +234,17 @@ func NewAccountMux(repo *data.Repository) func(r chi.Router) {
 				})
 
 				r.Post("/", func(w http.ResponseWriter, r *http.Request) {
+					callingAccountID, err := GetAccountIDFromRequestSessionCookie(repo, r)
+					if err != nil {
+						http.Error(w, fmt.Sprintf("getting account id: %v", err), http.StatusUnauthorized)
+						return
+					}
+					requestedAccountID := uuid.MustParse(chi.URLParam(r, "accountID"))
+					if callingAccountID != requestedAccountID {
+						http.Error(w, "deleting account", http.StatusUnauthorized)
+						return
+					}
+
 					r.ParseForm()
 					accountID := uuid.MustParse(chi.URLParam(r, "accountID"))
 					displayName := r.FormValue("displayName")
@@ -118,6 +257,17 @@ func NewAccountMux(repo *data.Repository) func(r chi.Router) {
 				})
 
 				r.Patch("/", func(w http.ResponseWriter, r *http.Request) {
+					callingAccountID, err := GetAccountIDFromRequestSessionCookie(repo, r)
+					if err != nil {
+						http.Error(w, fmt.Sprintf("getting account id: %v", err), http.StatusUnauthorized)
+						return
+					}
+					requestedAccountID := uuid.MustParse(chi.URLParam(r, "accountID"))
+					if callingAccountID != requestedAccountID {
+						http.Error(w, "deleting account", http.StatusUnauthorized)
+						return
+					}
+
 					r.ParseForm()
 					accountID := uuid.MustParse(chi.URLParam(r, "accountID"))
 					displayName := r.FormValue("displayName")
@@ -131,6 +281,17 @@ func NewAccountMux(repo *data.Repository) func(r chi.Router) {
 				})
 
 				r.Delete("/", func(w http.ResponseWriter, r *http.Request) {
+					callingAccountID, err := GetAccountIDFromRequestSessionCookie(repo, r)
+					if err != nil {
+						http.Error(w, fmt.Sprintf("getting account id: %v", err), http.StatusUnauthorized)
+						return
+					}
+					requestedAccountID := uuid.MustParse(chi.URLParam(r, "accountID"))
+					if callingAccountID != requestedAccountID {
+						http.Error(w, "deleting account", http.StatusUnauthorized)
+						return
+					}
+
 					accountID := uuid.MustParse(chi.URLParam(r, "accountID"))
 					if err := repo.DeleteProfile(r.Context(), accountID); err != nil {
 						http.Error(w, fmt.Sprintf("deleting profile: %v", err), http.StatusInternalServerError)
